requirements.txt
ollama>=0.3.0
feedparser>=6.0.11
python-dotenv>=1.0.0
twilio>=9.0.0
requests>=2.32.0
